+++ apps/prelude/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "prelude",
+    "name": "site",
     "version": "0.0.0",
     "private": true,
     "scripts": {
@@ -15,7 +15,6 @@
         "prop-types": "^15.8.1",
         "react": "18.2.0",
         "react-dom": "18.2.0",
-        "ui": "workspace:*",
         "zustand": "^4.3.6"
     },
     "devDependencies": {
