--- packages/ui/index.tsx
@@ -1,1 +1,0 @@
-export * from "./Button";
