--- packages/ui/package.json
@@ -1,27 +1,0 @@
-{
-    "name": "ui",
-    "version": "0.0.0",
-    "main": "./index.tsx",
-    "types": "./index.tsx",
-    "license": "MIT",
-    "scripts": {
-        "lint": "eslint *.ts*"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.20.12",
-        "@types/react": "^18.0.27",
-        "@types/react-dom": "^18.0.10",
-        "autoprefixer": "10.4.11",
-        "babel-loader": "^8.3.0",
-        "eslint": "^7.32.0",
-        "eslint-config-custom": "workspace:*",
-        "postcss": "^8.4.21",
-        "tailwindcss": "^3.2.6",
-        "tsconfig": "workspace:*",
-        "typescript": "^4.9.5"
-    },
-    "dependencies": {
-        "class-variance-authority": "^0.4.0",
-        "classnames": "^2.3.2"
-    }
-}
