--- packages/ui/tailwind.config.js
@@ -1,12 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-
-
-module.exports = {
-  content: [
-    "./app/**/*.{js,ts,jsx,tsx}",
-  ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
-}
