--- packages/ui/tsconfig.json
@@ -1,8 +1,0 @@
-{
-    "compilerOptions": {
-        "jsx": "react"
-    },
-    "extends": "tsconfig/react-library.json",
-    "include": [".", "../../tailwind.config.js"],
-    "exclude": ["dist", "build", "node_modules"]
-}
